+++ src/modules/file_index.js
@@ -93,14 +93,14 @@ export async function startCache() {
         if (isGood) {
           indexPath.set(`${source}/${node}`);
         } else {
-          logger.info(node, "filtered")
+          // logger.info(node, "filtered")
         }
         return isGood;
       },
     });
     // n = 1;
-    logger.info(`${source} Indexed`);
-    console.timeEnd("indexing " + source);
+    // logger.info(`${source} Indexed`);
+    console.timeEnd("Alpha | Indexing " + source);
   }
   logger.info(`Indexed: ${index.length} files. Depth: ${depthLimit}`);
   console.timeEnd("indexing");